Adding a more flexible way of defining conditions on rules.
Instead of trying to shove all the conditions logic in a single JSON
field, it makes more sense to treat each condition separately, then use
all the conditions together to build an evaluation tree, with conditions
as the nodes. This makes it possible to define conditions almost as
flexible as those you could make with Python code, but stored in the
database.