Split "selection" into a few separate classes.
This change better encapsulates the selection and manipulation of data
that is only available at the time the rule is checked.

Of particular note is the switch from passing objects and extra as
parameters, to passing an info object that simplifies the reuse of
deferred values, so they need only be calculated once per rule check.
(This was mostly designed with database queries in mind, to allow caching
of the query results in case they were needed in more than one condition.)

Also note the change from storing each condition as a row in the
database to storing the entire condition tree as a string. This could
potentially slow "compilation" of the rules, but if they are cached,
this chould not be too significant, compared to the muchsimpler queries
involved in retrieving and using rules.