Added a central store for continuations.
Also added a Rule base class to core.py and refined the way operators
work in conditions, in order to simplify the code a bit and get rid of
the _eval function and its big if statement.

Finally, added a discover function and associated setting
(RULES_MODULES) which will make sure that any required rules and
continuations are loaded before they are needed.