Tried to minimize the number of places maybe_const was called.
maybe_const was originally supposed to be an optimization, but it
quickly increased the code's complexity with a lot of "if isinstance"
statements. Now I've changed it so that certain places ALWAYS expect a
deferred value, but the get_value function self-optimizes the first time
it is called. This leaves the overhead of a function call, but then,
isinstance is a function call as well.