Made all deferred objects hashable.
Hashable objects can be used as keys instead of `id(self)` when caching
in `rules.deferred.Deferred._get_deferred_value`. This means that if two
selectors are equal, but different objects, the value will only be
cached once. Within a rule, there were already ways to handle that (the
with() statement at the beginning of the conditions), but now values
will be cached between multiple rules in a single evaluation.

Also made a slight improvement in maybe_const, using an exception
instead of returning a sentinel value.